rename instances of package_analysis to api_analysis also move the file api_analysis.dart handling the CLI to bin/
