+++ lib/src/api_analysis/common.dart
@@ -12,6 +12,7 @@ import 'package:analyzer/exception/exception.dart';
 import 'package:analyzer/file_system/file_system.dart' as resource;
 import 'package:collection/collection.dart';
 import 'package:http/http.dart' as http;
+import 'package:meta/meta.dart';
 import 'package:pana/pana.dart';
 import 'package:path/path.dart' as path;
 import 'package:pub_semver/pub_semver.dart';
@@ -152,6 +153,20 @@ abstract class PackageAnalysisContext {
     _targetPackageName = targetPackageName;
     this.packagePath =
         _dummy ? findPackagePath(targetPackageName!) : packagePath;
+  }
+}
+
+@internal
+class PackageAnalysisContextWithStderr extends PackageAnalysisContext {
+  PackageAnalysisContextWithStderr({
+    required super.session,
+    required super.packagePath,
+    super.targetPackageName,
+  });
+
+  @override
+  void warning(String message) {
+    stderr.writeln(message);
   }
 }
