+++ lib/src/api_analysis/kind.dart
