+++ lib/src/api_analysis/normalize_shape.dart
