+++ lib/src/api_analysis/shapes.dart
