+++ lib/src/api_analysis/shapes_ext.dart
