+++ test/api_analysis/testdata/lower_bound_constraint_issues/enum_values.yaml
