+++ test/api_analysis/testdata/lower_bound_constraint_issues/functions_methods.yaml
