+++ test/api_analysis/testdata/lower_bound_constraint_issues/private_class_exposed_by_target.yaml
