+++ test/api_analysis/testdata/lower_bound_constraint_issues/typedefs.yaml
