+++ test/api_analysis/testdata/summary/abstract_class.yaml
