+++ test/api_analysis/testdata/summary/classes_methods.yaml
