+++ test/api_analysis/testdata/summary/empty_package.yaml
