+++ test/api_analysis/testdata/summary/enums.yaml
