+++ test/api_analysis/testdata/summary/extensions.yaml
