+++ test/api_analysis/testdata/summary/inheritance.yaml
