+++ test/api_analysis/testdata/summary/show_hide.yaml
