+++ test/api_analysis/testdata/summary/typedefs.yaml
